moves rpc and protocol code to the core package, using agent as an entry point for nodes/executors/miners
