--- src/scripts/start_miner.py
@@ -1,3 +1,0 @@
-import miner
-
-miner.start()
