--- src/scripts/start_node.py
@@ -1,3 +1,0 @@
-import node
-
-node.start_node()
